move remaining random packages into src directory
