+++ bin/avs-cli/eigenda/prepare-registration.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/etherfi-protocol/etherfi-avs-operator-tool/keystore"
+	"github.com/etherfi-protocol/etherfi-avs-operator-tool/src/keystore"
 	"github.com/urfave/cli/v3"
 )
