+++ bin/avs-cli/eoracle/prepareRegistration.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 
 	"github.com/ethereum/go-ethereum/common"
-	"github.com/etherfi-protocol/etherfi-avs-operator-tool/keystore"
+	"github.com/etherfi-protocol/etherfi-avs-operator-tool/src/keystore"
 	"github.com/urfave/cli/v3"
 )
