+++ src/keystore/fixtures/fixture.bls.key.json
