+++ src/keystore/fixtures/fixture.bls.signature.json
