+++ src/keystore/fixtures/fixture.ecdsa.key.json
