+++ src/keystore/keystore_test.go
